+++ features/auth/index.ts
@@ -1,6 +1,3 @@
 export { User } from "./models/User";
 export type { UserDocument, UserWallet } from "./models/User";
-export { RegisterForm } from "./components/RegisterForm";
-export { LoginForm } from "./components/LoginForm";
 export { LogoutButton } from "./components/LogoutButton";
-export { FeeCommitment } from "./components/FeeCommitment";